producer/requirements.txt
kafka-python==2.0.2
faker==24.6.0
pandas==2.1.4
psycopg2-binary==2.9.9
numpy==1.26.2